==> spam_text_classification/4_test_predict.ipynb <==
import pandas as pd
from konlpy.tag import Mecab
from tensorflow.keras.preprocessing.text import Tokenizer

# 형태소 추출
spam_test['tokenized'] = spam_test.text.apply(mecab.morphs)

def join_arr(x):
    return ' '.join(x)

spam_test.tokenized = spam_test.tokenized.apply(join_arr)

# 토큰화, 벡터화
x_test = spam_test['tokenized'].values

tokenizer = Tokenizer()
x_test = tokenizer.texts_to_sequences(x_test)
x_test = pad_sequences(x_test, maxlen=max_len)

# 모델 불러오기
loaded_model = load_model('best_model.h5')

# 테스트데이터 예측 후 결과값 매핑
dl_predicted = loaded_model.predict(x_test)
dl_predicted = [1 if x>0.5 else 0 for x in dl_predicted]

# 결과값 df로 저장
submission = pd.DataFrame({'id' : spam_test.id, 'label' : dl_predicted1})

# label 텍스트로 변환
submission.label = submission.label.map({1 : 'spam', 0 : 'ham'})

# 결과값 내보내기
submission.to_csv("/aihub/data/submission.csv",index=False)
